Refactor: sensors
