+++ src/gui/map/map_editor.cpp
@@ -64,7 +64,7 @@
 #include "gui/widgets/symbol_widget.h"
 #include "gui/widgets/template_list_widget.h"
 #include "color_dock_widget.h"
-#include "compass.h"
+#include "sensors/compass.h"
 #include "fileformats/file_format_registry.h"
 #include "core/map.h"
 #include "map_dialog_rotate.h"
@@ -73,9 +73,9 @@
 #include "map_part_undo.h"
 #include "object_undo.h"
 #include "map_widget.h"
-#include "gps_display.h"
-#include "gps_temporary_markers.h"
-#include "gps_track_recorder.h"
+#include "sensors/gps_display.h"
+#include "sensors/gps_temporary_markers.h"
+#include "sensors/gps_track_recorder.h"
 #include "gui/main_window.h"
 #include "gui/print_widget.h"
 #include "gui/widgets/key_button_bar.h"
