+++ src/sensors/gps_temporary_markers.cpp
