+++ src/sensors/gps_track_recorder.h
